third_party/blink/web_tests/external/wpt/custom-elements/registries/Document-createElementNS-expected.txt

This is a testharness.js-based test.
[FAIL] createElementNS should use the specified scoped registry
  assert_true: expected true got false
[FAIL] createElementNS should create a builtin element regardless of a custom element registry specified
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] createElementNS should create an upgrade candidate when there is no matching definition in the specified registry
  assert_equals: expected "HTMLElement" but got "GlobalABElement"
[FAIL] createElementNS should create an upgrade candidate and the candidate should be upgraded when the element is defined
  assert_true: expected true got false
[FAIL] createElementNS on a non-HTML document should still handle registries correctly
  assert_equals: expected object "[object CustomElementRegistry]" but got null
Harness: the test ran to completion.
